Filter locally for remaining gcp resources.

+++ gcp/compute/instance_groups.go
@@ -2,12 +2,13 @@ package compute
 
 import (
 	"fmt"
+	"strings"
 
 	gcpcompute "google.golang.org/api/compute/v1"
 )
 
 type instanceGroupsClient interface {
-	ListInstanceGroups(zone, filter string) (*gcpcompute.InstanceGroupList, error)
+	ListInstanceGroups(zone string) (*gcpcompute.InstanceGroupList, error)
 	DeleteInstanceGroup(zone, instanceGroup string) error
 }
 
@@ -28,22 +29,26 @@ func NewInstanceGroups(client instanceGroupsClient, logger logger, zones map[str
 func (s InstanceGroups) Delete(filter string) error {
 	var groups []*gcpcompute.InstanceGroup
 	for _, zone := range s.zones {
-		l, err := s.client.ListInstanceGroups(zone, filter)
+		l, err := s.client.ListInstanceGroups(zone)
 		if err != nil {
 			return fmt.Errorf("Listing instance groups for zone %s: %s", zone, err)
 		}
 		groups = append(groups, l.Items...)
 	}
 
-	for _, i := range groups {
-		n := i.Name
+	for _, group := range groups {
+		n := group.Name
+
+		if !strings.Contains(n, filter) {
+			continue
+		}
 
 		proceed := s.logger.Prompt(fmt.Sprintf("Are you sure you want to delete instance group %s?", n))
 		if !proceed {
 			continue
 		}
 
-		zoneName := s.zones[i.Zone]
+		zoneName := s.zones[group.Zone]
 		if err := s.client.DeleteInstanceGroup(zoneName, n); err != nil {
 			s.logger.Printf("ERROR deleting instance group %s: %s\n", n, err)
 		} else {
